+++ bkg.config.json
@@ -1,0 +1,3 @@
+{
+    "entry": "client.js"
+}
