--- bun_dev/client.js
@@ -1,8 +1,0 @@
-const ws = new WebSocket("ws://localhost:3000/");
-
-ws.onmessage = (e) => {
-  console.log(e.data)
-};
-setInterval(() => {
-  ws.send("ping ");
-}, 800);
